+++ Source/BLE.Client/BLE.Client.Maui/Helpers/IPlatformHelpers.cs
@@ -1,0 +1,7 @@
+﻿namespace BLE.Client.Helpers
+{
+    public interface IPlatformHelpers
+    {
+        Task<PermissionStatus> CheckAndRequestBluetoothPermissions();
+    }
+}