+++ Source/BLE.Client/BLE.Client.Maui/Models/BLEDevices.cs
@@ -1,0 +1,10 @@
+﻿namespace BLE.Client.Maui.Models
+{
+    public class BLEDevices : List<BLEDevice>
+	{
+		public BLEDevices()
+		{
+		}
+	}
+}
+