+++ Source/BLE.Client/BLE.Client.Maui/Models/DeviceState.cs
@@ -1,0 +1,30 @@
+﻿/*using System;
+namespace BLE.Client.Maui.Models
+{
+
+    /// <summary>
+    /// Determines the connection state of the device.
+    /// </summary>
+	public enum DeviceState
+    {
+        /// <summary>
+        /// Device is disconnected.
+        /// </summary>
+        Disconnected,
+
+        /// <summary>
+        /// Device is connecting.
+        /// </summary>
+        Connecting,
+
+        /// <summary>
+        /// Device is connected.
+        /// </summary>
+        Connected,
+
+        /// <summary>
+        /// OnAndroid: Device is connected to the system. In order to use this device please call connect it by using the Adapter. 
+        /// </summary>
+        Limited
+    }
+}*/