+++ Source/BLE.Client/BLE.Client.Maui/Platforms/MacCatalyst/MacCatalystPlatformHelpers.cs
@@ -1,0 +1,16 @@
+﻿using System.Threading.Tasks;
+using BLE.Client.Helpers;
+
+namespace BLE.Client.Helpers
+{
+    public class MacCatalystPlatformHelpers : IPlatformHelpers
+    {
+        public Task<PermissionStatus> CheckAndRequestBluetoothPermissions()
+        {
+            System.Diagnostics.Debug.WriteLine("Into CheckAndRequestBluetoothPermissions ");
+
+            return Task.FromResult(PermissionStatus.Granted);
+        }
+    }
+}
+